Build system: style changes for cmake/modules/CheckDIRSymbolExists.cmake
These style changes mostly consisted of lowercasing the CMake command
names, but I also took this opportunity to improve the commentary
and output messages.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
by running cmake.  The relevant output from that run was

-- Looking for DIR symbol in sys/types.h;dirent.h
-- Looking for DIR symbol in sys/types.h;dirent.h - found

which is the expected result on that platform.