Build system: use check_symbol_exists rather than check_function_exists to find snprintf or _snprintf
This change works on Linux (tested below) and is also apparently
required for the VC14 (a.k.a Visual Studio 2015) platform because
certain io functions have been inlined for that platform.  See
comments for
<http://public.kitware.com/Bug/bug_relationship_graph.php?bug_id=15659>
which suggest that check_symbol_exists and not check_function_exists
should be used to check for the existence of functions that might be
inlined.