Build system: Use the official version of the freetype find module
Previously we used our own copy of the CMake-3.0.2 version of this
find module to benefit our cmake-2.8.x users, but this is pointless
now that our minimum version of CMake is 3.0.2 on Linux and higher on
other platforms.  Furthermore, this change means our users benefit
from the bug fixing in the official version of the find module that
has occurred for cmake versions > 3.0.2 if they happen to be using
such versions of cmake.