Build system: refine what gets dropped when -DDROP_GTK_PLUS_2_BUILDS=ON
The purpose of -DDROP_GTK_PLUS_2_BUILDS=ON is to drop GTK+ version 2
components from a GTK+ version 3 build on those platforms (e.g., where
epa_build has been used to provide GTK+ version 3 libraries but the
system libraries are still GTK+ version 2) where both GTK+ versions 2
and 3 are available.

This fix refines what is dropped.  Previously, anything having to do
with xcairo was dropped, e.g., the xcairo device and the
extXdrawable_demo application.  The extXdrawable_demo is indeed a pure
GTK+ version 2 application, but when GTK+ version 3 is available the
xcairo device is pure GTK+ version 3.  So for
-DDROP_GTK_PLUS_2_BUILDS=ON do not allow extXdrawable_demo to be built
but do allow the xcairo device to be built.