Build system: Disable OCAML_HAS_CAIRO
I have taken this step because substantial maintenance will be
required before this component of PLplot will configure, build, or run
again because of backwards incompatibilities that have been introduced
in ocaml support of cairo. For example, the package name "cairo2" no
longer exists any more and should probably be replaced by the package
name cairo, but that is just the start of the backwards
incompatibilities.