Build system: Complete rewrite of file and target dependency logic for test targets
The test issue that lead me to open this whole can of worms was that
wxPLViewer was not being built as a necessary dependency for
wxwidgets tests for ENABLE_DYNDRIVERS=OFF.  I was trying to use
target dependencies to force wxPLViewer to be built before the
wxwidgets target, but that idea obviously does not work correctly
for ENABLE_DYNDRIVERS=OFF since the wxwidgets target
does not exist for that case (i.e., the wxwidgets device driver
is part of the core plplot library in this case).

So the proper way to deal with this issue is to have wxwidgets
tests depend on the wxPLViewer target, but once I started reviewing
target dependencies of tests, I realized the CMake logic
implementing that had to be rewritten so that custom test targets
associated with interactive tests that normally do not have a
corresponding custom command should not have any DEPENDS and instead
add_dependencies should be used appropriately for that case.  The
single "virgin" tests below also showed a large number of target
dependency issues that had to be addressed.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux. I
performed "virgin" tests (i.e., where the test target was built alone
starting from an empty build tree without any other test target
obfuscating whether the particular single target dependencies are
properly built or not) on a subset of the noninteractive tests and on
essentially all interactive tests for both the ENABLE_DYNDRIVERS=ON
and ENABLE_DYNDRIVERS=OFF cases.  I also ran the following two
comprehensive tests:

scripts/comprehensive_test.sh --do_test_interactive no
scripts/comprehensive_test.sh --do_ctest no --do_test_noninteractive no

which is essentially equivalent to running

scripts/comprehensive_test.sh