Build system: Bump minimum CMake version to 3.3.2 for most platforms
The exceptions are the Linux or Cygwin platforms where the minimum
CMake version remains at 3.0.2 (to accomodate using system versions of
CMake for Cygwin and most modern Linux distributions).  In addition,
in the Linux or Cygwin case, the user is warned if they have a CMake
version between 3.2.0 and 3.3.1 because of a regression for that range
of CMake versions (corrected in 3.3.2) that compromised the find
ability of cmake.