Build system: Fix minor bug with the test_extXdrawable target
The issue was if extXdrawable_demo is not built for some reason (such
as the Cygwin package for gtk+ not correctly depending on the
libxrandr package) then the extXdrawable_demo target does not exist
and the test_extXdrawable target should not be built for this case.  A
simple change (adding AND TARGET extXdrawable_demo) to an if statement
controlling whether test_extXdrawable is built) fixed this issue, and
my thanks to Greg for running the Cygwin test that found this issue.