Experimental test Ada language support: add gcc to the list of compilers
When gcc is on the list (and for the case where ccache is installed)
the resulting Ada compiler is /usr/lib/ccache/gcc.  Previously I
thought that would be an issue, but evidentally (see test below) it is
not, i.e., the ccache result is the same as if /usr/bin/gcc (which is
a symlink to gcc-4.9) had been used for the compiler. (Note the
alternative Ada compiler name gnatgcc is also a symlink to gcc-4.9 on
Debian Jessie).

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux by
using the -DEXPERIMENTAL_SUPPORT=ON cmake option (which tests the
language support in cmake_experimental rather than cmake_working)
while otherwise following the test procedure documented in
cmake/test_ada/README.  For CMake-3.3.2 (built using the bootstrap
procedure) the run-time, valgrind, and ldd -r results were clean.