test_ada and PLplot Ada language support: restore explicit linking of Ada libraries to libgnat
This change solves an underlinking issue for the Ada libraries that
was nonconsequential on Linux (except that ldd -r on Ada libraries
revealed the issue) but which caused a build error on Cygwin.

On all platforms use the PRIVATE CMake keyword for the
target_link_libraries command used to link the Ada library with
libgnat.  That means that Ada applications that are linked to the Ada
library will not be overlinked to libgnat.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux using
CMake-3.3.2 (built using the bootstrap procedure).

For the test_ada project I followed the test procedure documented in
cmake/test_ada/README. The configuration with cmake, build, run-time,
valgrind, and ldd -r results (including ldd -r on the Ada library and
Ada executable) were all clean.

For the PLplot project I used the cmake options
-DDEFAULT_NO_DEVICES=ON -DPLD_ps=ON -DDEFAULT_NO_BINDINGS=ON
-DENABLE_ada=ON to limit the test to Ada and built the
test_noninteractive target in the build tree using VERBOSE=1 without
any obvious configuration, build, or run-time issues and with the
usual

Comparison test using psc device

ada
  Missing examples            :
  Differing graphical output  :  08 19
  Missing stdout              :
  Differing stdout            :
adathick
  Missing examples            :
  Differing graphical output  :  08 19
  Missing stdout              :
  Differing stdout            :

results.  In addition, the VERBOSE=1 build results revealed (as
expected from using the PRIVATE keyword) no underlinking of libgnat
for the Ada library and no overlinking of libgnat for any of the Ada
thin or thick examples. Furthermore, ldd -r gave good results on both
the Ada library and all Ada examples.