PLplot core library: remove octal number interpretation for numerical escape sequences
The long-standing but undocumented octal possibility for numerical
escape sequences (first introduced in 2008, but just brought to my
attention now) concerns the semantics of the strtoul C library
function (used in text2num which in turn is used in the chain of calls
from plptex, plptex3, plmtex, and plmtex3, but not plsym).

Previously the base in the strtoul call was set to 0.  Here is what
the strtoul Linux man page says about that case:

"If base is zero or 16, the string may then include a "0x" prefix, and
the number will be read in base 16; otherwise, a zero base is taken as
10 (decimal) unless the next character is '0', in which case it is
taken as 8 (octal)."

Furthermore, my experiments and also some of the man pages for strtoul
on the web (e.g., <http://illumos.org/man/3C/strtoul>) indicate the
Linux man page should state that strtoul interprets both "0x" and "0X"
as the hexadecimal case.  So for base 0, the strtoul interpretation of
numbers assumes a base following exactly what is done for C numerical
constants.

But that "feature" contradicts our documentation where hexadecimal and
decimal possibilities are documented (e.g., in
<http://plplot.sourceforge.net/docbook-manual/plplot-html-5.11.1/characters.html>
for the present release), but the octal possibility is undocumented.
Furthermore, newbie C programmers and programmers who have no
experience with C may be surprised by that octal interpretation.  For
example, before the present fix the Hershey encoding sequence
"#(0855)" (the one where I discovered this issue) chokes on the 8
because of the octal interpretation due to that leading zero, and such
users will experience even a larger surprise when they discover that,
e.g., "#(0123)" and "#(123)" silently give very different results
because of the respective octal (because of the leading zero) and
decimal interpretations.

Accordingly, instead of documenting this feature, I have made the
choice to remove the octal intepretation from our code (by setting
base for the strtoul call to 10 unless the string prefix is 0x or 0X
in which case base is set to 16).

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux using
some modified C examples where plptex and plmtex calls included
Hershey escape sequences with leading 0's (those numbers interpreted
now as decimal rather than the previous octal) and both leading 0x and
0X (those numbers interpreted as hexadecimal as always).