Update C example 7 to include a plptex Hershey escape sequence debugging mode
The introduced -plptex_mode example option replaces the plsym call by
the largely equivalent plptex call.  This change does not need to be
propagated to non-C languages. The intention of this debugging mode is
to allow direct comparison of plsym results (which use Hershey fonts
directly) with the corresponding plptex results for the equivalent
Hershey code.

For device drivers that use Hershey fonts (e.g., the xwin device) the
results with and without the -plptex_mode option are identical aside
from occasional horizontal offsets that typically are too small to be
seen but which for certain characters can be as large as 1/20th the
box width.  Despite those relatively large offsets in a minority of
cases, the only hypothesis I have to explain this difference is the
possibility of different internal coordinate systems and rounding of
those positional results.  But why are the large offsets only in X,
but not Y?  Anyhow, this is an issue that should be revisited once we
review and completely update the internal coordinate systems for
PLplot to not be limited to just 16-bit integers.

For unicode-aware devices (such as -dev qtwidget), there is a
translation done internally between Hershey escape sequences (as in
the present very extensive example when -plptex_mode is used) and
equivalent unicode.  And for -dev ps and -dev psc there is a further
translation between that derived unicode result and the corresponding
Type 1 Adobe font code for characters.  So to further debug the
Hershey-to-unicode transformation, it is planned to compare example 7
results with and without -plptex_mode for -dev qtwidget and to further
debug the Unicode to type1 transformation, the equivalent experimenet
is planned for -dev ps.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux by
comparing -dev xwin results for example 7 with and without the new
-plptex_mode option with results that have been described above.
Additional systematic testing using this mode is planned (also
described above).  One of those planned comparisons (for the -dev psc
device) has already been spot checked for the "#(855)" case, and as
currently discussed on the plplot-devel list, the plsym result shows
the Hershey font arrow while the plptex result shows a blank (unlike
other devices).  I am virtually positive this difference is due to a
missing entry in the Unicode to type1 transformation table, but it is
also probable this is a result of no glyph to represent an arrowhead
in the set of 35 standard type 1 fonts, i.e., this is an artifact of
the limitations of Type 1 fonts.