test_ada and PLplot Ada language support: adopt .o extension for object files for all platforms
Contrary to C++ (and other mainstream compilers) where this extension
is either .o for Unix or .obj otherwise we have evidence on both MinGW
(from years ago with a fix implemented in the old Ada language
support) and MinGW-w64 (now for the new Ada language support where
this fix is now required) that this extension is always .o for Ada
regardless of platform.  So use that assumption unless or until we
find a platform where that assumption is not correct (i.e., one where
the error message "gnat1: incorrect object file extension" is emitted
by the gnat1 Ada compiler front-end to the gcc compiler.