Fixed compile error when using no shapelib and deprecated is off and attempt to silence x19 warnings
The error was simply a variable defined in the wrong #ifdef section.

To avoid the warnings we need to not run the shapelib example when no
shapelib support is on. I tried to set up a PL_USE_SHAPEFILE, but this
doesn't seem to have worked. Awaiting Alan's input to fix.