Fortran binding: remove plplot_types.f90.in file that was added in error to the new binding
I added this file to the new binding because it was configured by
bindings/f95/CMakeLists.txt, but it defines a plplot_types module in
nameclash conflict with the plplot_types module defined in
plplot_bindings.f90 and the latter definition should be the module
that is used for the new binding Therefore remove this file as well as
the bindings/f95/CMakeLists.txt references to it.

N.B. The new Fortran binding still does not build, but the removal of
this confusion between two modules with identical names has
substantially reduced the build errors.