Fortran binding: Fix the last of the current set of build errors for libplplotf95
The fixes were to implement the definition of the maxlen parameter
(which is required by plparseopts) before the contains in the plplot
module, and the removal of "use plplot" from plparseopts which is one
of the files that are included in the plplot module beyond the
contains statement.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
by building the plplotf95 target.  Note, however, that the
test_diff_psc target still does not work because the Fortran
examples are linked with both the plplotf95 and plplotf95c libraries
and the latter does not exist (yet) in the new Fortran binding.