Fortran binding: Fix example build and run-time errors for the new binding
The fixes were the following:

* Name the C version of plparseopts correctly (as c_plparseopts and not f_plparseopts)

* Link the examples to both the plplotf95 and plplot libraries.  The
latter is required by the iso_c_binding method.  Very shortly (see
below) the plplotf95c library (the part of the Fortran interface that
is too complicated to be handled by the iso_c_binding method so it
must be written in C) will be added to the list of libraries that must
be linked.

* Temporarily drop the call to plparseopts in examplex/f95/x00f.f90.

This change is not needed to build that example, but it is needed to
run it.  Otherwise, the plparseopts call in this (and other examples)
currently just segfaults.  The issue is that plparseopts must interface an array
of character strings from fortran to C.  That complicated interfacing
task cannot be done by the iso_c_binding module and instead must be
done at the C level in the the planned small plplotf95c library that
will now be needed again.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
by building the ps and x00f95 targets and then running

valgrind examples/f95/x00f

(note no command line arguments are used because plparseopts is
temporarily not called in this example in order to get it to run
without segfaults.  The PostScript results looked good (after entering
the device and output file by hand), the valgrind report was pristine
(no errors, no leaks possible), and the ldd -r test gave clean results
as well.