Fortran binding: Initial implementation of libplplotf95c, the C part of the interface for the new binding
Some aspects of implementing the new Fortran binding cannot be handled
completely by the iso_c_binding Fortran module capabilities.  A good
example of this is the transformation of a blank-padded character
array Fortran argument to the required C array of pointers to
null-terminated character strings argument of plparsopts.  I have
dealt with this issue by implementing in the new plplotf95c C library
a C function to make the required argument transformation, and a
fc_plparseopts function that calls that transformation function and
then calls plparseopts with its argument in the correct (transformed)
form.

Because of this correct implementation of Fortran plparseopts
I was able to reinstate calling plparseopts for
examples/f95/x00f.f90, and test (see below) building and running
the Fortran examples (in the form they all had before the start of
this topic branch) with the new binding.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux by
attempting to build the test_diff_psc target.  Roughly ~60 per cent of
the examples did not build (typically because of certain parts of the
Fortran API that are currently missing from the new binding), but once
those examples were eliminated from the build list and run list, then
I could build the test_diff_psc target with the following result:

f95
  Missing examples            :  09 10 11 13 14 14a 15 16 17 18 19 20 21 22 23 28 29 31 33
  Differing graphical output  :  04 26
  Missing stdout              :
  Differing stdout            :

where the missing examples are the ones that do not yet build.  There
is obviously still quite a bit of work to do, but nevertheless the
examples that do build and have no differences with the equivalent C
results (i.e., 00 01 02 03 05 06 07 08 12 24 25 27 30) are a
promising start for the new Fortran binding, and the rest of this
project will simply consist of eliminating the problems that cause the
above differing graphical output and missing examples.