New Fortran binding: Drop installion of all PLplot module files other than plplot.mod
I discovered accidentally that our fortran examples (which only "use
plplot") can be built in the installed examples tree without access to
the other module files that are built in the build tree and which are
used by the plplot module to help build libplplotf95.  I assume this
is because the details of those other plplot modules that are use(d)
by the plplot module are made part of the plplot module.  Therefore,
following the principle of not exposing unneeded details to users, I
therefore dropped installing all plplot Fortran modules other than
plplot.mod.