New Fortran binding: Fix matrix_to_c bug and add API relevant to example 21
The matrix_to_c bug fix patch hunk was

     integer :: i_local

     allocate( carray(size(array,2),size(array,1)) )
-    allocate( caddress(size(array,2)) )
+    allocate( caddress(size(array,1)) )

     carray = transpose( array )

-    do i_local = 1,size(array,2)
+    do i_local = 1,size(array,1)
         caddress(i_local) = c_loc(carray(1,i_local))
     enddo
 end subroutine matrix_to_c

The point is that each column of the Fortran carray corresponds to a
row of the corresponding C array.  Furthermore, the C address of the
starting location of each column of the Fortran carray (or
equivalently the C address of the starting location of each row of the
corresponding C array) must be stored for later C use.  However, the
size of caddress is the number of columns of carray which is
size(array,1)) and not size(array,2)), and the do loop must be over
the range of 1 to size(array,1)) as well.

The bug in the previous code would have no effect if size(array,2) >
size(array,1) (other than the caddress array would be larger than
needed and its trailing (unused) values would be garbage.  However,
for the case where size(array,2) < size(array,1) the bug in the
previous code could have severe memory management effects because
the caddress array would be smaller than what is required by
the C routine.

The added API relevant to example 21 was plenv0, and plgriddata.

I also took the opportunity to change several

use iso_c_binding

instances that had crept back in to the appropriate

use iso_c_binding, only: ...

form.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by temporarily modifying examples/f95/x21.f90
to comment out the plshades call, building the x21f95 target,
and running it with

valgrind examples/f95/x21f -dev svg -fam -o testf.svg

That produced perfect (0 errors and no memory leaks are possible)
valgrind results and also plot file results (for the non-plshades page) that
agree exactly with the corresponding C pages.