Removed named mutex for perfomance boost to the wxWidgets driver.
As far as I can tell the mutex used in wxPLDevice::TransmitBuffer is not
needed as the buffer records the read and write positions and avoids
overtaking that way. Generating this mutex adds about 25% overhead on
Windows. Not sure about the overhead in Linux, but hopefuly this will
help performance issues reported by Alan.