New Fortran binding: Change naming conventions for certain entities
The plan for the new naming convention is the entities previously
called plrandd (a function with no arguments that returns a real), PLGraphicsIn (a
defined type which contains real quantities), and PLNOT_SET, PL_PI,
and PL_TWOPI (predefined real parameters) where the previous
associated real precision was plflt will be changed to always be
double precision regardless of the real precision of the PLplot C
library.  And the single-precision variants of these same quantities
will be defined with the names plrandd_single, PLGraphicsIn_single,
PL_NOTSET_single, PL_PI_single, and PL_TWOPI_single.

This commit implements this plan for plrandd and PLGraphicsIn.

I also took this opportunity to add the plgdrawmode API and to make
private_plflt, private_single, and private_double actually private.
These changes and the fact that private_plint was already actually
private required that the real(kind=private_plflt, private_single, or
private_double) or integer(kind=private_plint) type of result returned
by the plrandd and plgdrawmode functions had to have a type statement
in the body of the function rather than as part of the function
statement. To make that Fortran style uniform I used this form for
typing all function return values, and appended "!function type" to
the type statement for this case so it will be easy to find in the
future.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the plplotf95 library.