New Fortran binding: Add remaining API relevant to examples 16 (and 16a), 21, and 22
That is, I implemented the full plshade. plshades and plvect families
of routines as well as plstransform. For now, the latter required an
added dummy "data" argument to disambiguate the single- and
double-precision versions, but since that is a backwards-incompatible
change to the Fortran API, Alan will attempt to find another solution.

Furthermore, because the routines for the plstransform callback are used
by C they must be externally visible and have a unique name. Therefore
I put them in the file plplot_bindings instead of in the included file.

Tested by: Arjen Markus <arjen.markus@deltares.nl> on Cygwin with
gfortran as well as compiled using Intel Fortran under Windows. One
thing that remains to be done is plGetCursor.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the test_diff_psc target for the
shared library case.  The result summary was

f95
  Missing examples            :  19
  Differing graphical output  :  16
  Missing stdout              :
  Differing stdout            :

i.e., examples 16, 21, and 22 have now been added to the list of
Fortran examples which are buildable.

Examples 21 and 22 produce the same -dev psc results as the
corresponding C examples, but there are substantial differences for
example 16.  Further investigation showed the 16th Fortran example had
substantially fewer fills then the corresponding C example but those
extra C fills may have been cropped or behind other fills because the
C and Fortran results were visually indistinguishable. To simplify the
investigatio I tried nx, ny = 3, 4 for both the C and Fortran example,
but in that case there were identical results.  So the investigation
of this issue continues.

Example 16a was also added to the list of buildable examples, and the
resulting first page of the plot looks good (although it has been
years since I looked at it so I cannot guarantee there has been no
change).  The second page of the plot has some bad-looking vacant
areas, however, so clearly there is some issue with the plshade
variant that was used to produce that result.

The valgrind reports for

valgrind examples/f95/x<example id>f -dev svg -fam -o testf.svg

where example id is (unmodified or modified) 16, 16a, 21, and 22 are all perfect (0 errors, no
leaks are possible).