New Fortran binding: Add dimension checking, improve 2D transformation code consistency, and fix plimagefr bugs
The dimension checking additions (mostly for xg, and yg) were for
plshade_single_1, plshade_single_2, plshades_multiple_1,
plshades_multiple_2, plvectors_1, and plvectors_2.  This means we
should now have consistent dimension checking of xg and yg
for all variants of plcont, plshade, plshades, plvect, and plimagefr.

The transformation consistency improvements were to use identical code
for pltr2f for all cases where xg and yg were two-dimensional in the
new Fortran binding.  And to use identical code for transform for all
related signatures: _0, _1, _2, and _tr (referring to plcont, plshade,
plshades, plvect, and plimagefr signatures where xg, yg, and tr are
not in the argument list, 1D xg and yg are in the argument list but
not tr, 2D xg and yg are in the argument list but not tr, and xg, and
yg are not in the argument list, but tr is in the argument list.)

The plimagefr bug fix is I forgot to use xg_in and yg_in for the _1
and _2 cases.  This bug would generate run-time issues for the case
when kind=kp is different than private_plflt. But we don't test that
case now which explains why the plimagefr tests looked good before
this fix.

I also made sure all "use" statements had an associated "only" clause,
and I often checked that clause to make sure the minimum necessary was
being used from each module.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the test_diff_psc target for
the shared library case.  The fortran summary result remained the
same

f95
  Missing examples            :  19
  Differing graphical output  :  16
  Missing stdout              :
  Differing stdout            :

i.e., Fortran example 16 is still not consistent with the
corresponding C result.  And further looks at diff files between the
Fortran and C results for the svg case and visual inspection of
example 16a showed all issues were for 2D xg and yg and were generally
due to missing fills for the Fortran case.  And valgrind
results for Fortran examples 16 and 16a continue to be perfect.

So the above changes made absolutely no difference to the missing
fills issue for plshade and plshades with "_2" signature, but they
have cleared out a lot of obfuscating differences in code between the
various "_2" signatures for 2D xg and yg, and they now also do much
better checking of the dimensions of xg and yg in all cases.