Core library: move the warning concerning a NULL pltr callback from plcont to plfcont
plcont calls plfcont immediately with unchanged callback argument so
moving the warning to the latter function provides the warning for
both functions. Previously, the warning was only emitted for the
plcont call so a direct call to plfcont (which unconditionally
executes the callback) would previously have had run-time errors with
a NULL callback.