Changed strategy to get text size in wxWidgets driver
The delay in getting the text size from the wxPLViewer has been of
concern. Doing this required both the calling application and the viewer
to repeatedly poll the shared memory checking for updates which can
introduce significant delays. We now create a memory DC in the driver to
use specifically for getting text size when the user does not pass in a
wxDC.