Fix the interactive text wxBitmap at 24 bpp
Using the screen depth does not work on Linux as wxWidgets tries to
derive this from the top level window which doesn't exist. This causes a
crash. It doesn't actually matter what bit depth is used as the bitmap
is never actually drawn on - it just allows a valid wxMemoryDC to be
constructed for defining text size.