New Fortran binding: convert 22nd example to use modern Fortran idioms
As a result of this fairly minor change to this example (mostly to the
"transform" callback function) we now have a method for enforcing the
correct number and type of arguments for callback routines.  I assume
the essential here is that the transform callback now appears after
a "contains" statement which if I recall correctly automatically generates
an interface corresponding to transform, and because that interface
disagrees with the library one it generates a build-time error.  That,
of course is a vast improvement over the previous situation where the
only error for wrong number or type of arguments in the callback was
a segfault at run time.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the svg and x22f95 targets and then running

valgrind examples/f95/x22f -dev svg -fam -o testf.svg

This produced a perfect valgrind report (0 errors, no leaks are
possible) and also svg results that were identical to the
corresponding C example results.

I also made tests of the type safety of the callback by changing the
number and kind of arguments to transform.  In all these cases the gfortran
compiler emitted the following error message:

Error: There is no specific subroutine for the generic 'plstransform'

Strangely enough, despite gfortran being able to recognize when the
interface to the callback supplied by the user has the wrong number or
type of arguments, it still complains when attempting to build the
library that it cannot disambiguate plstransform if the disambiguation
parameter is consistently removed from the Fortran library
implemention of plstransform and also the use of plstransform in this
example.