Sort wxWidget initialization
In order to call GUI functions there appears to be a need for some
initialization which results from the creation of a wxApp. This seems to
be the case on Linux when creating wxFonts (which we need to be able to
do to get text sizes), although Windows seems less fussy. However, we
don't want to start the message loop,
so rather than using one of the wxWidget macros we just create and
initialize a default wxApp without calling wxApp::OnRun, which is the
function which begins the message loop.