In the wxWidgets driver only use wxGraphicContext under Windows
Ideally we want to use a wxGraphicsContext (wxGCDC) to determine the
size of text because that is what is used in wxPLViewer. However under
Linux creation of the wxGrahicsContext causes a crash. It might be that
this is only for wxGTK2, not wxGTK3 as wxWidgets does slightly different
things for these two cases, but until I have access to a GTK3 machine
for testing I can't be sure.
So for all OSs other than Windows we do the text measuring using a
wxMemoryDC. For Windows we use a wxGCDC.