Correct the interface routine plcontour_0 and implement a robust callback for transformations
The routine plfcont requires a transformation routine, this was missing from
the simplest version, plcontour_0. The current implementation uses the identity
mapping pltr0f.

A second small but essential change is the way the transformation callback
routines are handled. Because the Fortran standard does not allow
disambiguation via the argument list of procedure arguments, we decided to
use a different approach: the callback routine should always use double
precision, independent of the precision of the rest of the program

Tested by Arjen Markus <arjenmarkus@users.sourceforge.net> on Cygwin, using gfortran.