Restructured the wxWidgets text size again and remove tests for wxGraphicsContext
It seems that although the previous solution worked it is not supported
to create wxBitmaps or wxMemoryDCs in a console app. It is however
supported (although it seems undocumented) to create wxGraphicsContexts
from a wxImage in a console application - see
http://trac.wxwidgets.org/ticket/17319 and
http://trac.wxwidgets.org/ticket/17320

Also removed all checks for wxUSE_GRAPHICS_CONTEXT because this is now
supposedly set to 1 for all wxWidget platforms and also because there is
no officially supported fallback.