New Fortran binding: Add full data argument capability to plstransform and corresponding callback
I used the method given in "V: Using type(c_ptr) Pointer" of
<http://www.fortran90.org/src/best-practices.html> to allow the user
to pass an arbitrary user-defined definition as an argument of a
callback.  This additional signature for plstransform that allows this
is distinguished by an additional data argument and a Fortran callback
that has an additional data argument as well.  Furthermore, we have
done all the plumbing in plplot_bindings.f90 such that this callback
is interfaced properly with c_plstransform.

This method has been used in example 22 to pass what before was a
value stored in a common block as the data argument of the newly
defined transform_data Fortran callback that is used by plstransfrom
with a properly defined data argument in the example.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the svg and x22f95 targets and running

valgrind examples/f95/x22f -dev svg -fam -o testf.svg

with a perfect valgrind report and svg results that were identical
to the corresponding C results.