Core library: modify plmap code so that the library rather than examples emits warning messages concerning missing resources/capabilities
This change is a fix for the previous approach which used the
PL_USE_SHAPEFILES macro in examples to decide what warnings to emit, but
such a macro-based approach does not work for languages other than C
or C++ so we have now removed all references to PL_USE_SHAPEFILES
in the C and C++ examples (as well as the rest of our source tree).

Furthermore, I have now factored the src/plmap.c file so
that src/plmap.c is exclusively concerned with the shapelib
approach and therefore no longer has any reference to the old deprecated
plmap approach, and src/deprecated_plmap.c is exclusively concerned
with the old deprecated plmap approach and has no reference to
the modern shapelib approach.  The build system has been adjusted so
that when PL_DEPRECATED=OFF (the usual default) only the plmap.c code
and not deprecated_plmap.c is compiled into libplplot, and when PL_DEPRECATED=ON
only deprecated_plmap.c and not plmap.c is compiled into libplplot.

We discuss 3 cases:

I. PL_DEPRECATED=OFF, HAVE_SHAPELIB=ON
II. PL_DEPRECATED=OFF, HAVE_SHAPELIB=OFF
III. PL_DEPRECATED=ON

Case I.

This the usual case and it gives good results with no warning messages
for all pages of example 19.

Case II.

For this case most of the code in plmap.c is ignored by the
preprocessor (because the macro HAVE_SHAPELIB corresponding to the
CMake variable of the same name is not #defined).  Thus, the only code that
is left all the plmap* functions which
simply emit a warning message similar to

*** PLPLOT WARNING ***
plmap is a no-op because shapelib is not available.

and return.  Also, for what it is worth the plmeridians function is
available.

The net result for example 19 for Case II is as expected; all the
boxes are blank on each page of this example except for the
plmeridians results for two of them and "Baltimore" string for one of
them. Also a warning message similar to the above is emitted by the
library (not the example!) each time example 19 calls any of the
plmap* functions.

Case III.

For this case, the code in deprecated_plmap.c is used which makes no
reference to shapelib functionality or HAVE_SHAPELIB at all so you get
the same result regardless of whether HAVE_SHAPELIB is OFF or ON.  The
plmap function does plot the maps of the first 4 pages of example 19
using the old-fashioned non-shapelib method and data, but the rest of
the plmap* functions simply emit a warning similar to

*** PLPLOT WARNING ***
plmapfill not available when the -DPL_DEPRECATED=ON cmake option is used.

The net result for example 19 case III is as expected; the first four
pages of the example are plotted without issues or warning messages,
and the last shapelib-related page plots a blank box and the library
(not the example!) emits warning messages of the above type every time any plmap*
function is called in that example page. (Note, plmap itself is not
called by that page.)

Tested by: Alan W. Irwin ,airwin@users.sourceforge.net> on Linux
(Debian jessie).  For each of Cases I, II, and III, the expected
results as described above were obtained and the command

valgrind examples/c/x19c -dev svg -fam -o testc.svg

gave a perfect report (0 errors, no leaks are possible).